City Club App
This app was built for all supported iOS devices as a summer project on
behalf of my teacher and the City Club of San Diego. It allows users to
sign up for City Club events, view information about events, view past
videos of events, along with push notifications from the president to
its members.